requirements/base.txt
# Django
Django>=4.2.7,<4.2.8
djangorestframework==3.14.0
Pillow>=10.1.0,<10.2.0
djangorestframework-simplejwt==5.3.1
django-environ==0.11.2  # https://github.com/joke2k/django-environ
django-simple-history==3.4.0  # https://django-simple-history.readthedocs.io/en/latest/quick_start.html#configure
drf-yasg==1.21.7
django-extensions==3.2.3  # https://github.com/django-extensions/django-extensions
Werkzeug==3.0.1  # https://github.com/pallets/werkzeug

# Async Task
celery==5.3.6  # https://github.com/celery/celery
flower==2.0.1  # https://github.com/mher/flower

# Message broker
redis==5.0.1  # https://github.com/antirez/redis
